Added SIGALRM handling. Added timer unit test.
git-svn-id: svn://graphite/diminuto/trunk/Diminuto@34 5dfbf39a-1452-0410-8a97-9e58fedde8e3

+++ diminuto_delay.h
@@ -25,6 +25,6 @@
  * @return the number of microseconds remaining in the delay duration if
  * the function returned prematurely.
  */
-extern uint32_t diminuto_delay(uint32_t microseconds, int interruptable);
+extern uint64_t diminuto_delay(uint64_t microseconds, int interruptable);
 
 #endif
